+++ examples/compiler/library/scripts.build.ts
@@ -1,3 +1,3 @@
 import { libraryCompiler } from '@rockpack/compiler';
 
-void libraryCompiler('MyLib');
+void libraryCompiler({ name: 'MyLib' });