+++ packages/compiler/src/index.ts
@@ -23,6 +23,7 @@ export {
   getWebpack,
   // eslint-disable-next-line @typescript-eslint/no-deprecated -- the options overload is current; one overload is deprecated
   isomorphicCompiler,
+  // eslint-disable-next-line @typescript-eslint/no-deprecated -- the options overload is current; one overload is deprecated
   libraryCompiler,
   makeWebpackConfig,
   RockpackError,
@@ -30,4 +31,5 @@ export {
 };
 
 export type { IsomorphicCompilerOptions } from './compilers/isomorphic-compiler.js';
+export type { LibraryCompilerOptions } from './compilers/library-compiler.js';
 export type { RockpackErrorCode } from './errors/rockpack-error.js';